adding processing DAGs

+++ dags/historical_odds_ingestion.py
@@ -2,7 +2,7 @@ from airflow import DAG
 from airflow.operators.python import PythonOperator
 from airflow.exceptions import AirflowException
 
-from datetime import datetime, timedelta
+from datetime import timedelta
 import os
 import asyncio
 import aiohttp
@@ -18,8 +18,12 @@ load_dotenv()
 default_args = {
     "owner": "PENGUIN_DB",
     "depends_on_past": True,
-    "start_date": datetime(2024, 11, 29),  # NBA season 2025 start date is 2024-10-22
-    "end_date": datetime(2024, 12, 1),  # Today's date
+    "start_date": HistoricalConfig.SPORT_CONFIGS["nba"][
+        "start_date"
+    ],  # datetime(2024, 11, 29),  # NBA season 2025 start date is 2024-10-22
+    "end_date": HistoricalConfig.SPORT_CONFIGS["nba"][
+        "end_date"
+    ],  # datetime(2024, 12, 1),  # Today's date
     "email_on_failure": False,
     "retries": 0,
     "retry_delay": timedelta(minutes=5),
@@ -36,11 +40,10 @@ def fetch_odds_by_date(**context):
 
         # Get completed games for the date from previous games DAG
         games_data = context["task_instance"].xcom_pull(
-            dag_id="historical_games_ingestion",
-            task_ids="fetch_daily_games",
-            key="games_data",
+            dag_id="historical_nba_games_ingestion",
+            task_ids="fetch_daily_nba_games",
+            key="nba_games_data",
         )
-        print("GAMES DATA: ", games_data)
 
         if not games_data:
             print(f"No completed games found for date {date_str}")
@@ -67,39 +70,7 @@ def fetch_odds_by_date(**context):
 
                     return odds_data
                 return []
-        # for game in games_data:
-        #     if game["status"] in HistoricalConfig.FINISHED_STATUSES:
-        #         # Generate timestamps for game duration
-        #         timestamps = odds_connector.generate_game_timestamps(
-        #             game["date"],
-        #             HistoricalConfig.GAME_DURATIONS["nba"],
-        #             HistoricalConfig.ODDS_INTERVAL,
-        #         )
-        #
-        #         # Fetch odds for each timestamp
-        #         game_odds = []
-        #         async with aiohttp.ClientSession() as session:
-        #             for timestamp in timestamps:
-        #                 odds = await odds_connector.fetch_odds_snapshot(
-        #                     session, "nba", game["game_id"], timestamp
-        #                 )
-        #                 if odds:
-        #                     game_odds.append(
-        #                         {"timestamp": timestamp, "odds_data": odds}
-        #                     )
-        #
-        #         if game_odds:
-        #             all_odds_data.append(
-        #                 {"game_id": game["game_id"], "odds_history": game_odds}
-        #             )
-        #
-        # # Write to temporary location
-        # output_dir = f"/tmp/{HistoricalConfig.S3_PATHS['odds_prefix']}/{date_str}"
-        # os.makedirs(output_dir, exist_ok=True)
-        # with open(f"{output_dir}/odds.json", "w") as f:
-        #     json.dump(all_odds_data, f)
-        #
-        # return len(all_odds_data)
+
         except Exception as e:
             print(f"Error fetching odds: {str(e)}")
             raise AirflowException(f"Failed to fetch odds: {str(e)}")
@@ -112,6 +83,10 @@ def upload_to_s3_func(**context):
     local_path = f"/tmp/{HistoricalConfig.S3_PATHS['odds_prefix']}/{date_str}/odds.json"
     s3_path = f"{HistoricalConfig.S3_PATHS['odds_prefix']}/nba/{date_str}/odds.json"
 
+    if not os.path.exists(local_path):
+        print(f"No odds data file found for {date_str}, skipping upload")
+        return 0
+
     s3_client = boto3.client("s3")
     try:
         s3_client.upload_file(
@@ -120,6 +95,7 @@ def upload_to_s3_func(**context):
             s3_path,
             ExtraArgs={"ServerSideEncryption": "AES256"},
         )
+        return 1
     except Exception as e:
         print(f"Error uploading to S3: {str(e)}")
         raise
